fix: handle removed bucket on pr close

+++ src/actions/pr-closed.ts
@@ -7,6 +7,7 @@ import {
 import { s3Client } from '../clients/s3';
 import { validateEnvVars } from '../utils/env';
 import { deactivateDeployments, deleteDeployments } from '../utils/github';
+import { checkBucketExists } from '../utils/s3';
 
 export const requiredEnvVars = ['AWS_ACCESS_KEY_ID', 'AWS_SECRET_ACCESS_KEY'];
 
@@ -18,31 +19,34 @@ export const prClosed = async (
 
   validateEnvVars(requiredEnvVars);
 
-  console.log('Emptying S3 bucket...');
+  if (await checkBucketExists(bucketName)) {
+    console.log('Emptying S3 bucket...');
+    console.log('Fetching objects...');
 
-  console.log('Fetching objects...');
-
-  const objects = await s3Client.send(
-    new ListObjectsV2Command({ Bucket: bucketName })
-  );
-
-  if (objects.Contents && objects.Contents.length >= 1) {
-    console.log('Deleting objects...');
-    await s3Client.send(
-      new DeleteObjectsCommand({
-        Bucket: bucketName,
-        Delete: {
-          Objects: objects.Contents.map((object) => ({
-            Key: object.Key,
-          })),
-        },
-      })
+    const objects = await s3Client.send(
+      new ListObjectsV2Command({ Bucket: bucketName })
     );
-  } else {
-    console.log('S3 bucket already empty.');
-  }
 
-  await s3Client.send(new DeleteBucketCommand({ Bucket: bucketName }));
+    if (objects.Contents && objects.Contents.length >= 1) {
+      console.log('Deleting objects...');
+      await s3Client.send(
+        new DeleteObjectsCommand({
+          Bucket: bucketName,
+          Delete: {
+            Objects: objects.Contents.map((object) => ({
+              Key: object.Key,
+            })),
+          },
+        })
+      );
+    } else {
+      console.log('S3 bucket already empty.');
+    }
+
+    await s3Client.send(new DeleteBucketCommand({ Bucket: bucketName }));
+  } else {
+    console.log('S3 bucket does not exist.');
+  }
 
   await deactivateDeployments(repo, environmentPrefix);
   await deleteDeployments(repo, environmentPrefix);
